+++ lib/src/storage/v3/type_alias_declaration_storage_v3.g.dart
@@ -8,21 +8,21 @@ part of 'type_alias_declaration_storage_v3.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$_TypeAliasDeclarationStorageV3 _$$_TypeAliasDeclarationStorageV3FromJson(
-        Map<String, dynamic> json) =>
-    _$_TypeAliasDeclarationStorageV3(
-      name: json['name'] as String,
-      aliasedTypeName: json['aliasedTypeName'] as String,
-      isDeprecated: json['isDeprecated'] as bool,
-      isExperimental: json['isExperimental'] as bool,
-      entryPoints: (json['entryPoints'] as List<dynamic>)
-          .map((e) => e as String)
-          .toSet(),
-      relativePath: json['relativePath'] as String,
-    );
+_$TypeAliasDeclarationStorageV3Impl
+    _$$TypeAliasDeclarationStorageV3ImplFromJson(Map<String, dynamic> json) =>
+        _$TypeAliasDeclarationStorageV3Impl(
+          name: json['name'] as String,
+          aliasedTypeName: json['aliasedTypeName'] as String,
+          isDeprecated: json['isDeprecated'] as bool,
+          isExperimental: json['isExperimental'] as bool,
+          entryPoints: (json['entryPoints'] as List<dynamic>)
+              .map((e) => e as String)
+              .toSet(),
+          relativePath: json['relativePath'] as String,
+        );
 
-Map<String, dynamic> _$$_TypeAliasDeclarationStorageV3ToJson(
-        _$_TypeAliasDeclarationStorageV3 instance) =>
+Map<String, dynamic> _$$TypeAliasDeclarationStorageV3ImplToJson(
+        _$TypeAliasDeclarationStorageV3Impl instance) =>
     <String, dynamic>{
       'name': instance.name,
       'aliasedTypeName': instance.aliasedTypeName,