+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.16.3
 - adds more diff result reporting options (cli, json, markdown)
+- fixes broken local packages refs
 
 ## Version 0.16.2
 - fixes relative path handling in package config (leading to unresolvable types)